DRY null cipher
