+++ src/timesync/fuzz-nts.c
@@ -38,44 +38,3 @@ int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
 
         return 0;
 }
-
-/* null cipher */
-
-#define BLKSIZ 16
-
-int NTS_encrypt(uint8_t *ctxt,
-                const uint8_t *ptxt,
-                int ptxt_len,
-                const AssociatedData *info,
-                const struct NTS_AEADParam *nts,
-                const uint8_t *key) {
-        (void) info;
-        (void) nts;
-        (void) key;
-        memset(ctxt, 0xEE, BLKSIZ);
-        memmove(ctxt+BLKSIZ, ptxt, ptxt_len);
-        return ptxt_len + BLKSIZ;
-}
-
-int NTS_decrypt(uint8_t *ptxt,
-                const uint8_t *ctxt,
-                int ctxt_len,
-                const AssociatedData *info,
-                const struct NTS_AEADParam *nts,
-                const uint8_t *key) {
-        (void) info;
-        (void) nts;
-        (void) key;
-        if (ctxt_len < BLKSIZ)
-                return -1;
-
-        memmove(ptxt, ctxt+16, ctxt_len - BLKSIZ);
-        return ctxt_len - BLKSIZ;
-}
-
-const struct NTS_AEADParam* NTS_get_param(NTS_AEADAlgorithmType id) {
-        static struct NTS_AEADParam param = {
-                NTS_AEAD_AES_SIV_CMAC_256, 256/8, 16, 16, true, false, "AES-128-SIV"
-        };
-        return id? &param : NULL;
-}
